+++ src/main/java/com/homegarden/store/backend/utils/ProductFilterSpecification.java
@@ -5,6 +5,7 @@ import jakarta.persistence.criteria.Path;
 import jakarta.persistence.criteria.Predicate;
 import org.springframework.data.jpa.domain.Specification;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -12,8 +13,8 @@ public class ProductFilterSpecification {
 
     public static Specification<Product> filter(
             Long categoryId,
-            Double minPrice,
-            Double maxPrice,
+            BigDecimal minPrice,
+            BigDecimal maxPrice,
             Boolean discount,
             String sort
     ) {
